requirements.txt
streamlit==1.28.1
openai>=1.35.0
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
